fix(pdf): share text-layer extraction across platforms
